Reorganize admin file into folder
