+++ server/core/admin/utils.py
@@ -1,0 +1,42 @@
+from django.contrib import admin
+
+from .models import Program, StudentProfile, TeacherProfile
+
+
+class ActiveProgramFilter(admin.SimpleListFilter):
+    # Human-readable title which will be displayed in the
+    # right admin sidebar just above the filter options.
+    title = "program"
+
+    # Parameter for the filter that will be used in the URL query.
+    parameter_name = "program"
+
+    def lookups(self, request, model_admin):
+        """
+        Returns a list of tuples. The first element in each
+        tuple is the coded value for the option that will
+        appear in the URL query. The second element is the
+        human-readable name for the option that will appear
+        in the right sidebar.
+        """
+        active_programs = Program.get_active_programs()
+        return [(program.id, str(program)) for program in active_programs]
+
+    def queryset(self, request, queryset):
+        """
+        Returns the filtered queryset based on the value
+        provided in the query string and retrievable via
+        `self.value()`.
+        """
+        if self.value():
+            return queryset.filter(program=self.value())
+        else:
+            return queryset
+
+
+class StudentInline(admin.StackedInline):
+    model = StudentProfile
+
+
+class TeacherInline(admin.StackedInline):
+    model = TeacherProfile
